Add support for authorization and improve user interface and experience
This commit introduces a series of changes for better user experience and seamless navigation through our website application. User authorization was applied to all axios request headers, which improves the security and personalization of the app. It also changes the server URL to localhost, allowing developers to test API endpoints seamlessly.
For each user role, the proper interface is presented. If the user is an administrator, they will have an access to manage users view.
Throughout different views, such as ad posting and fetching, headers were added and localStorage is used for persisting user's data throughout sessions. This significantly improves the user experience, as the user can now timely receive personalized content and options.
Small UI adjustments have also been applied, for instance, the color of a backend button was changed in AdPostPage.vue. It was necessary for the consistency of the design and comprehensiveness of the user interface.
The new files AdminLogin.vue and Manage.vue have been added to produce administrative functionalities: login for admins and manage users view correspondingly.
Overall, these updates create a more secure, user-friendly, and appealing interaction with our application.