Remove unused files and refactor code
Files `src/components/Footer.vue` and `vue.config.js` were deleted because they are no longer needed in the project.

In `tsconfig.json`, paths were added in the include and exclude properties and the "allowJs" compiler option was set to true for better project structure and management.

Progress indicators were added in the `AdPostPage.vue` and `Login.vue` to show the loading status whenever a network request is made.

In `Login.vue`, session data was cleaned up, renamed 'AdminPremission' to 'AdminPermission' and flags were changed to string boolean representations.

New libraries `NProgress` and `view-ui-plus` were added in the `package-lock.json` for better UI and control.

The code was also cleaned and beautified in several places.