Add naive-ui and search functionality in MainPage.vue
Added naive-ui to the project where it was defined as a package in the package-lock.json file for some UI enhancements. Also implemented a search feature in the MainPage.vue. Users can now search the service or posts by entering the search keyword in the search bar. This feature fetches posts and services dynamically from the back-end and populates them accordingly. Pagination of search results was also implemented to limit display on the page.

Minor changes were made in Manage.vue to handle better pagination and permissions.