Add direct chat feature from ad post to user
Integrated a new feature which enables users to directly initiate chat with ad poster from the MainPage. Replaced the username symbol in MainPage component with a clickable phone icon, routing to chatroom with the specific username as a query in the url. Updated ChatRoom to fetch target user's data on mount using this username from the url, improving user experience by direct chat initiation.